+++ lib/widget/opacity.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/cupertino.dart';
+
+extension WidgetOpacityX on Widget {
+  Opacity opacity({
+    Key? key,
+    required double opacity,
+    bool alwaysIncludeSemantics = false,
+  }) =>
+      Opacity(
+        key: key,
+        opacity: opacity,
+        alwaysIncludeSemantics: alwaysIncludeSemantics,
+        child: this,
+      );
+}